reAct_agent/requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-community
langchain-google-genai
langgraph
python-dotenv
pydantic
chromadb
sentence-transformers
IPython 